requirements.txt
Scrapy
psycopg2-binary
SQLAlchemy
python-dotenv
lxml
